move and rename 'get-credentials' to 'src/lib/shared/authentiaction-lib.mjs'; also added tags for stack create
