+++ src/lib/actions/destroy.mjs
@@ -3,7 +3,7 @@ import { CloudFormationClient, DeleteStackCommand } from '@aws-sdk/client-cloudf
 
 import { emptyBucket } from 's3-empty-bucket'
 
-import { getCredentials } from './lib/get-credentials'
+import { getCredentials } from '../shared/authentication-lib'
 import { progressLogger } from '../shared/progress-logger'
 import { SiteTemplate } from '../shared/site-template'
 import { trackStackStatus } from './lib/track-stack-status'
