+++ src/lib/actions/ensure-ssl-certificate.mjs
@@ -3,7 +3,7 @@ import { ACMClient, RequestCertificateCommand } from '@aws-sdk/client-acm'
 import { Questioner } from 'question-and-answer'
 
 import { findCertificate } from './lib/find-certificate'
-import { getCredentials } from './lib/get-credentials'
+import { getCredentials } from '../shared/authentication-lib'
 import { progressLogger } from '../shared/progress-logger'
 
 const ensureSSLCertificate = async ({ apexDomain, db, globalOptions, siteInfo }) => {
