+++ src/lib/shared/authentication-lib.mjs
